Switches to Emotion for nav and search.
