+++ components/navigation/nav-item.tsx
@@ -1,21 +1,5 @@
 import { IconButton, Tooltip } from '@mui/material';
-import { styled } from '@mui/material/styles';
 import Link from 'next/link';
-
-const PREFIX = 'NavItem';
-
-const classes = {
-	inactive: `${ PREFIX }-inactive`,
-};
-
-const StyledLink = styled( Link )( ( { theme } ) => ( {
-	[ `& .${ classes.inactive }` ]: {
-		color: theme.palette.primary.light,
-		'&:hover': {
-			color: 'inherit',
-		},
-	},
-} ) );
 
 type NavItemLinkProps = {
 	name: string;
@@ -37,17 +21,22 @@ const NavItem: React.FC< NavItemLinkProps > = ( {
 		active = true;
 	}
 	return (
-		<StyledLink href={ href }>
+		<Link href={ href }>
 			<Tooltip title={ name }>
 				<IconButton
 					aria-label={ name }
-					className={ active ? '' : classes.inactive }
+					sx={ [
+						active && {
+							color: 'primary.light',
+							'&:hover': { color: 'inherit' },
+						},
+					] }
 					size="large"
 				>
 					{ icon }
 				</IconButton>
 			</Tooltip>
-		</StyledLink>
+		</Link>
 	);
 };
