+++ components/search-bar/types.d.ts
@@ -1,0 +1,11 @@
+export type SearchBarProps = {
+	minimized?: boolean;
+	inputId?: string;
+	options: SearchOption[];
+};
+
+export type SearchOption = {
+	title: string;
+	href: string;
+	type?: string;
+};
